requirements.txt
torch==2.7.1+cu118
torchvision==0.22.1+cu118
numpy
scipy
pillow
joblib
scikit-learn
pandas
matplotlib
seaborn 